% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsimx_wrapper_options.R
\name{apsimx_wrapper_options}
\alias{apsimx_wrapper_options}
\title{Getting a apsimx_wrapper options list with initialized fields}
\usage{
apsimx_wrapper_options(apsimx_path, apsimx_file, ...)
}
\arguments{
\item{apsimx_path}{Path of the apsimx binary executable file (Models.exe)}

\item{apsimx_file}{Path to the .apsimx file to be run}

\item{predicted_table_name}{Name of the predicted table in the datastore}

\item{observed_table_name}{Name of the observed table in the datastore}

\item{time_display}{Logical value used to display (TRUE) or not (FALSE)
simulations duration}

\item{multi_process}{Logical value used to run (TRUE) or not (FALSE) apsim
in multi-process mode. This will generally result in faster execution times
on systems with many cores (ie in a HPC/cluster environment).}
}
\value{
A list containing apsimx model apsimx_wrapper options
}
\description{
This function returns a default options list
}
\examples{

}
