% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsimx_wrapper.R
\name{apsimx_wrapper}
\alias{apsimx_wrapper}
\title{Running .apsimx files from txt input files stored in one directory
per `situation`, simulated results are returned in a list}
\usage{
apsimx_wrapper(param_values = NULL, sit_var_dates_mask = NULL,
  prior_information = NULL, model_options)
}
\arguments{
\item{param_values}{named vector containing the value(s) and names of the
parameters to force (optional)}

\item{sit_var_dates_mask}{List of situations, variables and dates for which
simulated values should be returned. Typically a list containing the
observations to which simulations should be compared as provided by
apsimxRFiles::read_obs_to_list}

\item{prior_information}{Prior information on the parameters to estimate.
For the moment only uniform distribution are allowed.
Either a list containing (named) vectors of upper and lower
bounds (\code{ub} and \code{lb}), or a named list containing for each
parameter the list of situations per group (\code{sit_list})
and the vector of upper and lower bounds (one value per group) (\code{ub} and \code{lb})}

\item{model_options}{List containing any information needed by the model.
In the case of apsimx: \code{apsimx_path} the path of apsimx executable file and
\code{apsimx_file} the path of the directory containing the apsimx input data
for each USM (one folder per USM where apsimx input files are stored in txt
format)}
}
\value{
A list containing simulated values (\code{sim_list}) and a flag
(\code{flag_allsim}) indicating if all required situations, variables and
dates were simulated.
}
\description{
This function uses ApsimX directly through a system call, can
force ApsimX input parameters with values given in arguments.
}
\examples{

}
